Pages folder refactoring and auth middleware
