+++ package.json
@@ -14,7 +14,6 @@
     "bcrypt": "^5.0.0",
     "body-parser": "^1.19.0",
     "formik": "^2.1.5",
-    "google-maps-react": "^2.0.6",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.10.2",
     "multer": "^1.4.2",
