+++ pages/api/formsAchievement.js
@@ -2,6 +2,7 @@ import nextConnect from "next-connect";
 import Achivement from "../../models/achievement";
 import Gallery from "../../models/gallery";
 import mongooseConnection from "../../middleware/database";
+import auth from '../../middleware/auth';
 
 const multer = require("multer");
 const storage = multer.memoryStorage();
@@ -53,6 +54,7 @@ function onError(err, req, res) {
 }
 
 const handler = nextConnect({ onError });
+handler.use(auth);
 
 handler.use(upload.fields([{ name: "images" }, { name: "pdfs" }]));
